requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
decord==0.6.0
deepspeed==0.16.3
dominate>=2.4.0
einops==0.8.0
fancycompleter==0.9.1
filelock==3.17.0
fsspec==2025.2.0
hjson==3.1.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pdbpp==0.10.3
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyrepl==0.9.0
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
sympy==1.13.1
visdom>=0.1.8.8
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
-e git+https://github.com/huggingface/transformers/@c772bff31a65c9c6002d0e74797cb130959a3716#egg=transformers
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wandb
wheel==0.45.1
wmctrl==0.5
opencv-python
openai





